+++ tests/GameFinder.Wine.Tests/Bottles/Test_BottlesWinePrefixProperties.cs
@@ -9,7 +9,10 @@ public partial class BottlesTests
     [Theory, AutoFileSystem]
     public void Test_ShouldWork_GetBottlesConfigurationFile(AbsolutePath prefixDirectory)
     {
-        var bottleWinePrefix = new BottlesWinePrefix(prefixDirectory);
+        var bottleWinePrefix = new BottlesWinePrefix
+        {
+            ConfigurationDirectory = prefixDirectory,
+        };
         bottleWinePrefix.GetBottlesConfigFile().Should().Be(prefixDirectory.CombineUnchecked("bottle.yml"));
     }
 }